Decrease `Duration` range from `uint64_t` to `uint32_t`
The `uint64_t` was chosen based on `std::chrono::milliseconds`. However, we use SDL for time related values, which uses `Uint32`.

The SDL2-Mixer library uses `int` values for millisecond timings. So not quite the same as the main SDL library, but closer than `uint64_t`. Considering the values are used for fade-in and fade-out times, it wouldn't make sense to use negative values in that context.

+++ NAS2D/Duration.h
@@ -5,5 +5,5 @@
 
 struct Duration
 {
-	uint64_t milliseconds;
+	uint32_t milliseconds;
 };
